#137 [feat] 레이아웃 모듈화
