+++ src/index.css
@@ -33,7 +33,7 @@ a:hover {
 html {
   height: 100%;
   zoom: 1; /* 기본 줌 설정 */
-  overflow: hidden;
+  /* overflow: hidden; */
 }
 
 body {
@@ -42,7 +42,7 @@ body {
   height: 100%;
   min-width: 320px;
   min-height: 100vh;
-  overflow: hidden;
+  /* overflow: hidden; */
 }
 
 input,
